extract menu items to hook
